server/pyproject.toml
[tool.poetry]
name = "rag-python"
version = "0.1.0"
description = "Backend service to interact with an AI RAG"
authors = ["Michael Bailey <mbailey1302@gmail.com>"]
license = "Propietary"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.10,<3.12"
structlog = "^23.2.0"
python-jose = "^3.3.0"
uvicorn = "^0.24.0.post1"
gunicorn = "^21.2.0"
beanie = "^1.23.6"
asgi-correlation-id = "^4.2.0"
passlib = {version = "^1.7.4", extras = ["bcrypt"]}
pydantic = {version = "^2.5.0", extras = ["email"]}
pydantic-settings = "^2.1.0"
python-multipart = "^0.0.6"
fastapi = "^0.104.1"
redis = "^5.0.1"
pymilvus = "^2.3.3"
llama-index = "^0.9.3.post1"
transformers = "^4.35.2"
s3fs = "^2023.10.0"
llama-cpp-python = "^0.2.19"
torch = "^2.1.1"
boto3-stubs = {version = "^1.29.6", extras = ["s3"]}
pypdf = "^3.17.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
isort = "^5.12.0"
flake8 = "^6.0.0"
faker = "^17.6.0"
black = "^23.3.0"
mypy = "^1.2.0"
httpx = "^0.24.1"
moto = "^4.2.9"

[tool.isort]
profile="black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
